backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
boto3==1.42.59
botocore==1.42.59
langgraph==1.0.10
langchain-aws==1.3.1
python-jose[cryptography]==3.5.0
httpx==0.28.1
anyio==4.12.1
